--- src/components/Bookings.js
@@ -1,21 +1,0 @@
-import React from "react";
-import Search from "./Search.js";
-import SearchResults from "./SearchResults.js";
-import FakeBookings from "./data/fakeBookings.json";
-
-const Bookings = () => {
-  const search = searchVal => {
-    console.info("TO DO!", searchVal);
-  };
-
-  return (
-    <div className="App-content">
-      <div className="container">
-        <Search search={search} />
-        <SearchResults result={FakeBookings} />
-      </div>
-    </div>
-  );
-};
-
-export default Bookings;
