+++ src/components/Bookings.jsx
@@ -1,0 +1,22 @@
+import React, { useState } from "react";
+import Search from "../Search.js";
+import SearchResults from "./SearchResults.jsx";
+import FakeBookings from "../data/fakeBookings.json";
+
+const Bookings = () => {
+  const search = searchVal => {
+    console.info("TO DO!", searchVal);
+  };
+  const [bookings, setBookings] = useState(FakeBookings);
+
+  return (
+    <div className="App-content">
+      <div className="container">
+        <Search search={search} />
+        <SearchResults result={bookings} />
+      </div>
+    </div>
+  );
+};
+
+export default Bookings;
