+++ src/components/RestaurantButton.jsx
@@ -1,0 +1,7 @@
+import React from "react";
+
+const RestaurantButton = props => {
+  return <button className="btn btn-primary" onClick={props.orderOne} />;
+};
+
+export default RestaurantButton;
